Library restructuring.
git-svn-id: http://ellcc.org/svn/ellcc/branches/reset@2212 82857169-6e38-49d7-aaa8-07fb558f12fc

+++ libecc/src/linux/write.c
@@ -1,7 +1,7 @@
-#include <sysdep.h>
+#include <syscall.h>
 #include <unistd.h>
 
-int write(int fd, const void* buf, size_t count)
+ssize_t write(int fd, const void* buf, size_t count)
 {
     return INLINE_SYSCALL(write, 3, fd, buf, count);
 }
